etc: Fix int to pointer typecasting build warnings
This is a bug from the khronos vulkan header file. It is scheduled to be resolved in
milestone "P1". See the following issue page.

https://github.com/KhronosGroup/Vulkan-LoaderAndValidationLayers/issues/36

When we use typecasting to pointer type, we should use uintptr_t type. it provides
safe pointer typecasting on 32bit and 64bit. We can revert this commit when the
above issue is resolved.

Change-Id: Id53d17c13c0dafc341889875140469c0f2943e67