swapchain: Signal fence and semaphore in vkAcquireNextImageKHR()
Users should wait for fence or semaphore to be signaled in order to
use the acquired VkImage. We signal them immediately because the
dequeued buffer is already ready to be used. Later, we can do some
kinda lazy signaling by returning buffer index immediately and signal
fence or semaphore when wl_buffer.release actually arrives.

Change-Id: I59c555f939834448ce4251552452c4a1bc54f26f